requirements.txt
numpy==1.19.1
pillow==7.2.0
ruamel.yaml==0.16.10
torch==1.5.0
torchvision==0.6.0
tensorboard==2.2.1

